Add tvm.testing.random_data helper

+++ topi/tests/python/test_topi_bitserial_conv2d.py
@@ -24,7 +24,7 @@ from tvm.contrib.pickle_memoize import memoize
 def generate_quantized_np(shape, bits, out_dtype):
     min_val = 0
     max_val = 1 << bits
-    return np.random.randint(min_val, max_val, size=shape).astype(out_dtype)
+    return np.random.randint(min_val, max_val, size=shape, dtype=out_dtype)
 
 def verify_bitserial_conv2d_nchw(batch, in_size, in_channel, num_filter, kernel, stride, padding,
                                  activation_bits, weight_bits, unipolar):
